+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "comfyui-jakeupgrade"
 description = "ComfyUI workflow customization by Jake."
-version = "2.3.8"
+version = "2.3.9"
 license = {file = "LICENSE"}
 dependencies = ["opencv-python", "piexif", "torch", "torchvision", "numpy", "pillow", "simpleeval", "functools", "PyYAML", "toml"]
